solve slice functions merge conflict #REF457
